requirements.txt
requests
python-dotenv
websockets
streamlit
matplotlib
pandas
